+++ Source/DotSpatial.Data/Feature.cs
@@ -249,7 +249,7 @@ namespace DotSpatial.Data
                 {
                     return ShapeIndex?.RecordNumber - 1 ?? -2; // -1 because RecordNumber for shapefiles is 1-based.
 
-                    // todo: The better will be remove RecordNumber from public interface to avoid ±1 issues.
+                    // todo: The better will be remove RecordNumber from public interface to avoid Â±1 issues.
                 }
 
                 return _parentFeatureSet.Features.IndexOf(this);
@@ -451,6 +451,31 @@ namespace DotSpatial.Data
                     break;
                 case OgcGeometryType.MultiPoint:
                     featureType = FeatureType.MultiPoint;
+                    break;
+                case OgcGeometryType.GeometryCollection:
+                    IGeometryCollection geomCollection = geometry as IGeometryCollection;
+                    if (geomCollection != null)
+                    {
+                        // Check to see if every featureType in the GeometryCollection matches
+                        // else leave as FeatureType.Unspecified
+                        FeatureType testFeatureType = FeatureType.Unspecified;
+                        for (int i = 0; i < geomCollection.Count; i++)
+                        {
+                            if (i == 0)
+                            {
+                                testFeatureType = FeatureTypeFromGeometryType(geomCollection[i]);
+                            }
+                            else
+                            {
+                                FeatureType tempFeatureType = FeatureTypeFromGeometryType(geomCollection[i]);
+                                if (testFeatureType != tempFeatureType)
+                                    return FeatureType.Unspecified; // if feature types do not match, then return Unspecified
+                            }
+                        }
+
+                        featureType = testFeatureType;
+                    }
+
                     break;
             }